requirements.txt
openai
ipython 
python-dotenv
jupyter
markitdown